+++ examples/example-umami/components/meta.tsx
@@ -19,9 +19,7 @@ export function Meta({ title, description }: MetaProps) {
         rel="canonical"
         href={absoluteURL(router.asPath !== "/" ? router.asPath : "")}
       />
-      <title>
-        {`${title} | ${siteConfig.name}`}
-      </title>
+      <title>{`${title} | ${siteConfig.name}`}</title>
       <meta name="description" content={description || siteConfig.slogan} />
       <meta
         property="og:image"