+++ packages/next-drupal/src/client.ts
@@ -1,3 +1,7 @@
+import { Jsona } from "jsona"
+import { stringify } from "qs"
+import { JsonApiErrors } from "./jsonapi-errors"
+import { logger as defaultLogger } from "./logger"
 import type {
   GetStaticPathsContext,
   GetStaticPathsResult,
@@ -5,42 +9,39 @@ import type {
   NextApiRequest,
   NextApiResponse,
 } from "next"
-import { stringify } from "qs"
-import Jsona from "jsona"
-
 import type {
-  JsonApiResource,
-  Locale,
   AccessToken,
-  JsonApiResponse,
-  JsonApiWithLocaleOptions,
-  JsonApiParams,
-  DrupalTranslatedPath,
-  DrupalMenuLinkContent,
-  FetchOptions,
-  DrupalClientOptions,
   BaseUrl,
-  JsonApiWithAuthOptions,
-  PathPrefix,
-  JsonApiResourceWithPath,
-  PathAlias,
-  PreviewOptions,
-  JsonApiWithCacheOptions,
-  JsonApiCreateResourceBody,
-  JsonApiUpdateResourceBody,
-  DrupalClientAuthUsernamePassword,
   DrupalClientAuthAccessToken,
   DrupalClientAuthClientIdSecret,
-  JsonApiCreateFileResourceBody,
-  DrupalView,
+  DrupalClientAuthUsernamePassword,
+  DrupalClientOptions,
   DrupalFile,
+  DrupalMenuLinkContent,
+  DrupalTranslatedPath,
+  DrupalView,
+  FetchOptions,
+  JsonApiCreateFileResourceBody,
+  JsonApiCreateResourceBody,
+  JsonApiParams,
+  JsonApiResource,
+  JsonApiResourceWithPath,
+  JsonApiResponse,
+  JsonApiUpdateResourceBody,
+  JsonApiWithAuthOptions,
+  JsonApiWithCacheOptions,
+  JsonApiWithLocaleOptions,
+  Locale,
+  PathAlias,
+  PathPrefix,
 } from "./types"
-import { logger as defaultLogger } from "./logger"
-import { JsonApiErrors } from "./jsonapi-errors"
 
 const DEFAULT_API_PREFIX = "/jsonapi"
 const DEFAULT_FRONT_PAGE = "/home"
 const DEFAULT_WITH_AUTH = false
+export const DRAFT_DATA_COOKIE_NAME = "draftData"
+// See https://vercel.com/docs/workflow-collaboration/draft-mode
+export const DRAFT_MODE_COOKIE_NAME = "__prerender_bypass"
 
 // From simple_oauth.
 const DEFAULT_AUTH_URL = "/oauth/token"
@@ -78,9 +79,9 @@ function isClientIdSecretAuth(
 export class DrupalClient {
   baseUrl: BaseUrl
 
-  debug: DrupalClientOptions["debug"]
-
   frontPage: DrupalClientOptions["frontPage"]
+
+  private isDebugEnabled: DrupalClientOptions["debug"]
 
   private serializer: DrupalClientOptions["serializer"]
 
@@ -112,8 +113,6 @@ export class DrupalClient {
 
   private previewSecret?: DrupalClientOptions["previewSecret"]
 
-  private forceIframeSameSiteCookie?: DrupalClientOptions["forceIframeSameSiteCookie"]
-
   /**
    * Instantiates a new DrupalClient.
    *
@@ -141,7 +140,6 @@ export class DrupalClient {
       auth,
       previewSecret,
       accessToken,
-      forceIframeSameSiteCookie = false,
       throwJsonApiErrors = true,
     } = options
 
@@ -149,7 +147,7 @@ export class DrupalClient {
     this.apiPrefix = apiPrefix
     this.serializer = serializer
     this.frontPage = frontPage
-    this.debug = debug
+    this.isDebugEnabled = !!debug
     this.useDefaultResourceTypeEntry = useDefaultResourceTypeEntry
     this.fetcher = fetcher
     this.auth = auth
@@ -159,7 +157,6 @@ export class DrupalClient {
     this.previewSecret = previewSecret
     this.cache = cache
     this.accessToken = accessToken
-    this.forceIframeSameSiteCookie = forceIframeSameSiteCookie
     this.throwJsonApiErrors = throwJsonApiErrors
 
     // Do not throw errors in production.
@@ -167,7 +164,7 @@ export class DrupalClient {
       this.throwJsonApiErrors = false
     }
 
-    this._debug("Debug mode is on.")
+    this.debug("Debug mode is on.")
   }
 
   set apiPrefix(apiPrefix: DrupalClientOptions["apiPrefix"]) {
@@ -216,7 +213,6 @@ export class DrupalClient {
     this.tokenExpiresOn = Date.now() + token.expires_in * 1000
   }
 
-  /* eslint-disable @typescript-eslint/no-explicit-any */
   async fetch(input: RequestInfo, init?: FetchOptions): Promise<Response> {
     init = {
       ...init,
@@ -230,7 +226,7 @@ export class DrupalClient {
     // Using the auth set on the client.
     // TODO: Abstract this to a re-usable.
     if (init?.withAuth) {
-      this._debug(`Using authenticated request.`)
+      this.debug(`Using authenticated request.`)
 
       if (init.withAuth === true) {
         if (typeof this._auth === "undefined") {
@@ -242,15 +238,15 @@ export class DrupalClient {
         // By default, if withAuth is set to true, we use the auth configured
         // in the client constructor.
         if (typeof this._auth === "function") {
-          this._debug(`Using custom auth callback.`)
+          this.debug(`Using custom auth callback.`)
 
           init["headers"]["Authorization"] = this._auth()
         } else if (typeof this._auth === "string") {
-          this._debug(`Using custom authorization header.`)
+          this.debug(`Using custom authorization header.`)
 
           init["headers"]["Authorization"] = this._auth
         } else if (typeof this._auth === "object") {
-          this._debug(`Using custom auth credentials.`)
+          this.debug(`Using custom auth credentials.`)
 
           if (isBasicAuth(this._auth)) {
             const basic = Buffer.from(
@@ -260,7 +256,7 @@ export class DrupalClient {
             init["headers"]["Authorization"] = `Basic ${basic}`
           } else if (isClientIdSecretAuth(this._auth)) {
             // Use the built-in client_credentials grant.
-            this._debug(`Using default auth (client_credentials).`)
+            this.debug(`Using default auth (client_credentials).`)
 
             // Fetch an access token and add it to the request.
             // Access token can be fetched from cache or using a custom auth method.
@@ -269,21 +265,20 @@ export class DrupalClient {
               init["headers"]["Authorization"] = `Bearer ${token.access_token}`
             }
           } else if (isAccessTokenAuth(this._auth)) {
-            init["headers"][
-              "Authorization"
-            ] = `${this._auth.token_type} ${this._auth.access_token}`
+            init["headers"]["Authorization"] =
+              `${this._auth.token_type} ${this._auth.access_token}`
           }
         }
       } else if (typeof init.withAuth === "string") {
-        this._debug(`Using custom authorization header.`)
+        this.debug(`Using custom authorization header.`)
 
         init["headers"]["Authorization"] = init.withAuth
       } else if (typeof init.withAuth === "function") {
-        this._debug(`Using custom authorization callback.`)
+        this.debug(`Using custom authorization callback.`)
 
         init["headers"]["Authorization"] = init.withAuth()
       } else if (isBasicAuth(init.withAuth)) {
-        this._debug(`Using basic authorization header`)
+        this.debug(`Using basic authorization header.`)
 
         const basic = Buffer.from(
           `${init.withAuth.username}:${init.withAuth.password}`
@@ -298,19 +293,18 @@ export class DrupalClient {
           init["headers"]["Authorization"] = `Bearer ${token.access_token}`
         }
       } else if (isAccessTokenAuth(init.withAuth)) {
-        init["headers"][
-          "Authorization"
-        ] = `${init.withAuth.token_type} ${init.withAuth.access_token}`
+        init["headers"]["Authorization"] =
+          `${init.withAuth.token_type} ${init.withAuth.access_token}`
       }
     }
 
     if (this.fetcher) {
-      this._debug(`Using custom fetcher.`)
+      this.debug(`Using custom fetcher, fetching: ${input}`)
 
       return await this.fetcher(input, init)
     }
 
-    this._debug(`Using default fetch (polyfilled by Next.js).`)
+    this.debug(`Using default fetch, fetching: ${input}`)
 
     return await fetch(input, init)
   }
@@ -333,8 +327,7 @@ export class DrupalClient {
 
     const url = this.buildUrl(apiPath, options?.params)
 
-    this._debug(`Creating resource of type ${type}.`)
-    this._debug(url.toString())
+    this.debug(`Creating resource of type ${type}.`)
 
     // Add type to body.
     body.data.type = type
@@ -377,8 +370,7 @@ export class DrupalClient {
       options?.params
     )
 
-    this._debug(`Creating file resource for media of type ${type}.`)
-    this._debug(url.toString())
+    this.debug(`Creating file resource for media of type ${type}.`)
 
     const response = await this.fetch(url.toString(), {
       method: "POST",
@@ -419,8 +411,7 @@ export class DrupalClient {
 
     const url = this.buildUrl(`${apiPath}/${uuid}`, options?.params)
 
-    this._debug(`Updating resource of type ${type} with id ${uuid}.`)
-    this._debug(url.toString())
+    this.debug(`Updating resource of type ${type} with id ${uuid}.`)
 
     // Update body.
     body.data.type = type
@@ -459,8 +450,7 @@ export class DrupalClient {
 
     const url = this.buildUrl(`${apiPath}/${uuid}`, options?.params)
 
-    this._debug(`Deleting resource of type ${type} with id ${uuid}.`)
-    this._debug(url.toString())
+    this.debug(`Deleting resource of type ${type} with id ${uuid}.`)
 
     const response = await this.fetch(url.toString(), {
       method: "DELETE",
@@ -493,7 +483,7 @@ export class DrupalClient {
       const cached = (await this.cache.get(options.cacheKey)) as string
 
       if (cached) {
-        this._debug(`Returning cached resource ${type} with id ${uuid}`)
+        this.debug(`Returning cached resource ${type} with id ${uuid}.`)
 
         const json = JSON.parse(cached)
 
@@ -508,8 +498,7 @@ export class DrupalClient {
 
     const url = this.buildUrl(`${apiPath}/${uuid}`, options?.params)
 
-    this._debug(`Fetching resource ${type} with id ${uuid}.`)
-    this._debug(url.toString())
+    this.debug(`Fetching resource ${type} with id ${uuid}.`)
 
     const response = await this.fetch(url.toString(), {
       withAuth: options.withAuth,
@@ -691,6 +680,8 @@ export class DrupalClient {
       _format: "json",
     })
 
+    this.debug(`Fetching resource by path, ${path}.`)
+
     const response = await this.fetch(url.toString(), {
       method: "POST",
       credentials: "include",
@@ -743,8 +734,7 @@ export class DrupalClient {
       ...options?.params,
     })
 
-    this._debug(`Fetching resource collection of type ${type}`)
-    this._debug(url.toString())
+    this.debug(`Fetching resource collection of type ${type}.`)
 
     const response = await this.fetch(url.toString(), {
       withAuth: options.withAuth,
@@ -914,6 +904,8 @@ export class DrupalClient {
       path,
     })
 
+    this.debug(`Fetching translated path, ${path}.`)
+
     const response = await this.fetch(url.toString(), {
       withAuth: options.withAuth,
     })
@@ -994,6 +986,8 @@ export class DrupalClient {
     )
 
     try {
+      this.debug(`Fetching JSON:API index.`)
+
       const response = await this.fetch(url.toString(), {
         // As per https://www.drupal.org/node/2984034 /jsonapi is public.
         withAuth: false,
@@ -1048,62 +1042,120 @@ export class DrupalClient {
     return href
   }
 
+  async validateDraftUrl(searchParams: URLSearchParams): Promise<Response> {
+    const slug = searchParams.get("slug")
+
+    this.debug(`Fetching draft url validation for ${slug}.`)
+
+    // Fetch the headless CMS to check if the provided `slug` exists
+    let response: Response
+    try {
+      // Validate the draft url.
+      const validateUrl = this.buildUrl("/next/draft-url").toString()
+      response = await this.fetch(validateUrl, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(Object.fromEntries(searchParams.entries())),
+      })
+    } catch (error) {
+      response = new Response(JSON.stringify({ message: error.message }), {
+        status: 401,
+      })
+    }
+
+    this.debug(
+      response.status !== 200
+        ? `Could not validate slug, ${slug}`
+        : `Validated slug, ${slug}`
+    )
+
+    return response
+  }
+
   async preview(
-    request?: NextApiRequest,
-    response?: NextApiResponse,
-    options?: PreviewOptions
+    request: NextApiRequest,
+    response: NextApiResponse,
+    options?: Parameters<NextApiResponse["setDraftMode"]>[0]
   ) {
-    const { slug, resourceVersion, plugin } = request.query
+    const { slug, resourceVersion, plugin, secret, scope, ...draftData } =
+      request.query
+    const useDraftMode = options?.enable
 
     try {
       // Always clear preview data to handle different scopes.
       response.clearPreviewData()
 
       // Validate the preview url.
-      const validateUrl = this.buildUrl("/next/preview-url")
-      const result = await this.fetch(validateUrl.toString(), {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(request.query),
-      })
-
-      if (!result.ok) {
-        response.statusCode = result.status
-
-        return response.json(await result.json())
-      }
+      const result = await this.validateDraftUrl(
+        new URL(request.url, `http://${request.headers.host}`).searchParams
+      )
 
       const validationPayload = await result.json()
-
-      response.setPreviewData({
+      const previewData = {
         resourceVersion,
         plugin,
         ...validationPayload,
-      })
+      }
+
+      if (!result.ok) {
+        this.debug(`Draft url validation error: ${validationPayload.message}`)
+        response.statusCode = result.status
+        return response.json(validationPayload)
+      }
+
+      // Optionally turn on draft mode.
+      if (useDraftMode) {
+        response.setDraftMode(options)
+      }
+
+      // Turns on preview mode and adds preview data to Next.js' static context.
+      response.setPreviewData(previewData)
 
       // Fix issue with cookie.
       // See https://github.com/vercel/next.js/discussions/32238.
       // See https://github.com/vercel/next.js/blob/d895a50abbc8f91726daa2d7ebc22c58f58aabbb/packages/next/server/api-utils/node.ts#L504.
-      if (this.forceIframeSameSiteCookie) {
-        const previous = response.getHeader("Set-Cookie") as string[]
-        previous.forEach((cookie, index) => {
-          previous[index] = cookie.replace(
-            "SameSite=Lax",
-            "SameSite=None;Secure"
-          )
-        })
-        response.setHeader(`Set-Cookie`, previous)
+      const cookies = (response.getHeader("Set-Cookie") as string[]).map(
+        (cookie) => cookie.replace("SameSite=Lax", "SameSite=None; Secure")
+      )
+      if (useDraftMode) {
+        // Adds preview data for use in app router pages.
+        cookies.push(
+          `${DRAFT_DATA_COOKIE_NAME}=${encodeURIComponent(
+            JSON.stringify({ slug, resourceVersion, ...draftData })
+          )}; Path=/; HttpOnly; SameSite=None; Secure`
+        )
       }
+      response.setHeader("Set-Cookie", cookies)
 
-      // We can safely redirect to the slug since this has been validated on the server.
+      // We can safely redirect to the slug since this has been validated on the
+      // server.
       response.writeHead(307, { Location: slug })
+
+      this.debug(`${useDraftMode ? "Draft" : "Preview"} mode enabled.`)
 
       return response.end()
     } catch (error) {
+      this.debug(`Preview failed: ${error.message}`)
       return response.status(422).end()
     }
+  }
+
+  async previewDisable(request: NextApiRequest, response: NextApiResponse) {
+    // Disable both preview and draft modes.
+    response.clearPreviewData()
+    response.setDraftMode({ enable: false })
+
+    // Delete the draft data cookie.
+    const cookies = response.getHeader("Set-Cookie") as string[]
+    cookies.push(
+      `${DRAFT_DATA_COOKIE_NAME}=; Path=/; Expires=Thu, 01 Jan 1970 00:00:00 GMT; HttpOnly; SameSite=None; Secure`
+    )
+    response.setHeader("Set-Cookie", cookies)
+
+    response.writeHead(307, { Location: "/" })
+    response.end()
   }
 
   async getMenu<T = DrupalMenuLinkContent>(
@@ -1127,7 +1179,7 @@ export class DrupalClient {
       const cached = (await this.cache.get(options.cacheKey)) as string
 
       if (cached) {
-        this._debug(`Returning cached menu items for ${name}`)
+        this.debug(`Returning cached menu items for ${name}.`)
         return JSON.parse(cached)
       }
     }
@@ -1142,8 +1194,7 @@ export class DrupalClient {
       options.params
     )
 
-    this._debug(`Fetching menu items for ${name}.`)
-    this._debug(url.toString())
+    this.debug(`Fetching menu items for ${name}.`)
 
     const response = await this.fetch(url.toString(), {
       withAuth: options.withAuth,
@@ -1216,6 +1267,8 @@ export class DrupalClient {
       options.params
     )
 
+    this.debug(`Fetching view, ${viewId}.${displayId}.`)
+
     const response = await this.fetch(url.toString(), {
       withAuth: options.withAuth,
     })
@@ -1255,6 +1308,8 @@ export class DrupalClient {
       `${localePrefix}${this.apiPrefix}/index/${name}`,
       options.params
     )
+
+    this.debug(`Fetching search index, ${name}.`)
 
     const response = await this.fetch(url.toString(), {
       withAuth: options.withAuth,
@@ -1334,11 +1389,11 @@ export class DrupalClient {
       this._token &&
       Date.now() < this.tokenExpiresOn
     ) {
-      this._debug(`Using existing access token.`)
+      this.debug(`Using existing access token.`)
       return this._token
     }
 
-    this._debug(`Fetching new access token.`)
+    this.debug(`Fetching new access token.`)
 
     const basic = Buffer.from(`${clientId}:${clientSecret}`).toString("base64")
 
@@ -1347,7 +1402,7 @@ export class DrupalClient {
     if (opts?.scope) {
       body = `${body}&scope=${opts.scope}`
 
-      this._debug(`Using scope: ${opts.scope}`)
+      this.debug(`Using scope: ${opts.scope}`)
     }
 
     const response = await this.fetch(url.toString(), {
@@ -1365,8 +1420,6 @@ export class DrupalClient {
     }
 
     const result: AccessToken = await response.json()
-
-    this._debug(result)
 
     this.token = result
 
@@ -1415,13 +1468,13 @@ export class DrupalClient {
     return message
   }
 
-  private _debug(message) {
-    !!this.debug && this.logger.debug(message)
+  debug(message) {
+    this.isDebugEnabled && this.logger.debug(message)
   }
 
   // Error handling.
-  // If throwErrors is enable, we show errors in the Next.js overlay.
-  // Otherwise we log the errors even if debugging is turned off.
+  // If throwErrors is enabled, we show errors in the Next.js overlay.
+  // Otherwise, we log the errors even if debugging is turned off.
   // In production, errors are always logged never thrown.
   private throwError(error: Error) {
     if (!this.throwJsonApiErrors) {