+++ packages/next-drupal/src/jsonapi-errors.ts
@@ -1,4 +1,4 @@
-import { JsonApiError } from "./types"
+import type { JsonApiError } from "./types"
 
 export class JsonApiErrors extends Error {
   errors: JsonApiError[] | string