+++ starters/basic-starter/lib/drupal.ts
@@ -1,8 +1,14 @@
 import { DrupalClient } from "next-drupal"
 
-export const drupal = new DrupalClient(
-  process.env.NEXT_PUBLIC_DRUPAL_BASE_URL,
-  {
-    previewSecret: process.env.DRUPAL_PREVIEW_SECRET,
-  }
-)
+const baseUrl: string = process.env.NEXT_PUBLIC_DRUPAL_BASE_URL || ""
+const clientId = process.env.DRUPAL_CLIENT_ID || ""
+const clientSecret = process.env.DRUPAL_CLIENT_SECRET || ""
+const previewSecret = process.env.DRUPAL_PREVIEW_SECRET
+
+export const drupal = new DrupalClient(baseUrl, {
+  auth: {
+    clientId,
+    clientSecret,
+  },
+  previewSecret,
+})