+++ www/config/tutorials.ts
@@ -1,14 +1,26 @@
 export const tutorialsConfig = [
   {
     title: "Quick Start",
+    excerpt: "Get Started with Next.js for Drupal using JSON:API.",
+    url: "/learn/quick-start",
     items: [],
   },
   {
     title: "Quick Start (GraphQL)",
+    excerpt: "Get Started with Next.js for Drupal using GraphQL.",
+    url: "/learn/graphql",
     items: [],
   },
   {
     title: "Preview Mode",
+    excerpt: "Implement Preview Mode for Entity Types.",
+    url: "/learn/preview-mode",
+    items: [],
+  },
+  {
+    title: "On-demand Revalidation",
+    excerpt: "Implement On-demand Revalidation",
+    url: "/learn/on-demand-revalidation",
     items: [],
   },
 ]