+++ examples/example-umami/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0](https://github.com/chapter-three/next-drupal/compare/example-umami@1.1.4...example-umami@1.2.0) (2022-12-06)
+
+**Note:** Version bump only for package example-umami
+
+
+
+
+
 ## [1.1.4](https://github.com/chapter-three/next-drupal/compare/example-umami@1.1.3...example-umami@1.1.4) (2022-12-06)