+++ starters/graphql-starter/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0](https://github.com/chapter-three/next-drupal/compare/graphql-starter@0.0.2...graphql-starter@0.1.0) (2022-12-06)
+
+**Note:** Version bump only for package graphql-starter
+
+
+
+
+
 ## 0.0.2 (2022-12-06)